feat: support horizontal widgets in XML generation Refactor XML generation into specific domain components to group and render widgets in horizontal rows.
