+++ cv-image-to-xml/src/main/java/org/appdevforall/codeonthego/computervision/domain/model/ScaledBox.kt
@@ -1,0 +1,8 @@
+package org.appdevforall.codeonthego.computervision.domain.model
+
+import android.graphics.Rect
+
+data class ScaledBox(
+    val label: String, var text: String, val x: Int, val y: Int, val w: Int, val h: Int,
+    val centerX: Int, val centerY: Int, val rect: Rect
+)
